revert Fleet generation
