+++ resources/META-INF/plugin-java.xml
@@ -22,10 +22,6 @@
       </action>
       <action id="grammar.Generate.ParserUtil" class="org.intellij.grammar.actions.BnfGenerateParserUtilAction"/>
       <action id="grammar.Generate.JFlexLexer" class="org.intellij.grammar.actions.BnfGenerateLexerAction"/>
-      <separator/>
-      <action id="grammar.fleet.Generate" class="org.intellij.grammar.fleet.GenerateFleetAction"/>
-      <action id="grammar.fleet.Generate.JFlexLexer" class="org.intellij.grammar.fleet.BnfGenerateFleetLexerAction"/>
-      <action id="grammar.fleet.Run.JFlex" class="org.intellij.grammar.fleet.BnfRunFleetJFlexAction"/>
       <add-to-group group-id="grammar.file.group" anchor="first"/>
     </group>
   </actions>
