+++ resources/messages/GrammarKitBundle.properties
@@ -14,9 +14,6 @@ action.grammar.Generate.JFlexLexer.text=Generate JFlex Lexer
 action.grammar.Generate.ParserUtil.text=Generate Parser Util
 action.grammar.Run.JFlex.text=Run JFlex Generator
 action.grammar.Generate.text=Generate Parser Code
-action.grammar.fleet.Generate.text=Fleet: Generate Parser Code
-action.grammar.fleet.Generate.JFlexLexer.text=Fleet: Generate JFlex Lexer
-action.grammar.fleet.Run.JFlex.text=Fleet: Run JFlex Generator
 
 intention.category.bnf=Grammar-Kit BNF
 intention.flip.arguments.text=Flip arguments
