+++ src/org/intellij/grammar/generator/ParserGenerator.java
@@ -10,6 +10,7 @@ import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.ui.MessageType;
 import com.intellij.openapi.util.Pair;
 import com.intellij.openapi.util.Trinity;
+import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.CommonClassNames;
 import com.intellij.psi.NavigatablePsiElement;
@@ -30,7 +31,10 @@ import org.intellij.grammar.psi.impl.GrammarUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.*;
@@ -54,7 +58,7 @@ import static org.intellij.grammar.psi.BnfTypes.*;
  * @author gregory
  *         Date 16.07.11 10:41
  */
-public class ParserGenerator extends GeneratorBase {
+public class ParserGenerator {
   public static final Logger LOG = Logger.getInstance(ParserGenerator.class);
 
   static class RuleInfo {
@@ -116,12 +120,26 @@ public class ParserGenerator extends GeneratorBase {
   private final Set<String> myTokensUsedInGrammar = new LinkedHashSet<>();
   private final boolean myNoStubs;
 
-  protected final String myParserUtilClass;
-  protected final String myPsiImplUtilClass;
+  private final BnfFile myFile;
+  private final String mySourcePath;
+  private final String myOutputPath;
+  private final String myPackagePrefix;
+  private final String myGrammarRoot;
+  private final String myGrammarRootParser;
+  private final String myParserUtilClass;
+  private final String myPsiImplUtilClass;
   private final String myPsiTreeUtilClass;
 
-  protected final String myVisitorClassName;
-  protected final String myTypeHolderClass;
+  private final NameFormat myIntfClassFormat;
+  private final NameFormat myImplClassFormat;
+
+  private final String myVisitorClassName;
+  private final String myTypeHolderClass;
+
+
+  private int myOffset;
+  private PrintWriter myOut;
+  private NameShortener myShortener;
 
   private final RuleGraphHelper myGraphHelper;
   private final ExpressionHelper myExpressionHelper;
@@ -130,16 +148,23 @@ public class ParserGenerator extends GeneratorBase {
   private final JavaHelper myJavaHelper;
 
   protected final Names N;
+  protected final GenOptions G;
   protected final IntelliJPlatformConstants C;
 
   public ParserGenerator(@NotNull BnfFile psiFile,
                          @NotNull String sourcePath,
                          @NotNull String outputPath,
                          @NotNull String packagePrefix) {
-    super(psiFile, sourcePath, outputPath, packagePrefix);
+    myFile = psiFile;
+    mySourcePath = sourcePath;
+    myOutputPath = outputPath;
+    myPackagePrefix = packagePrefix;
 
+    G = new GenOptions(myFile);
     N = G.names;
 
+    myIntfClassFormat = getPsiClassFormat(myFile);
+    myImplClassFormat = getPsiImplClassFormat(myFile);
     myParserUtilClass = getRootAttribute(myFile, KnownAttribute.PARSER_UTIL_CLASS);
     myPsiImplUtilClass = getRootAttribute(myFile, KnownAttribute.PSI_IMPL_UTIL_CLASS);
     myPsiTreeUtilClass = getRootAttribute(myFile, KnownAttribute.PSI_TREE_UTIL_CLASS);
@@ -161,6 +186,8 @@ public class ParserGenerator extends GeneratorBase {
     myJavaHelper = JavaHelper.getJavaHelper(myFile);
 
     List<BnfRule> rules = psiFile.getRules();
+    BnfRule rootRule = rules.isEmpty() ? null : rules.get(0);
+    myGrammarRoot = rootRule == null ? null : rootRule.getName();
     for (BnfRule r : rules) {
       String ruleName = r.getName();
       boolean noPsi = !hasPsiClass(r);
@@ -172,6 +199,7 @@ public class ParserGenerator extends GeneratorBase {
         noPsi ? null : getRulePsiClassName(r, myIntfClassFormat), noPsi ? null : getRulePsiClassName(r, myImplClassFormat),
         noPsi ? null : getAttribute(r, KnownAttribute.MIXIN), noPsi ? null : getAttribute(r, KnownAttribute.STUB_CLASS)));
     }
+    myGrammarRootParser = rootRule == null ? null : ruleInfo(rootRule).parserClass;
     myNoStubs = JBIterable.from(myRuleInfos.values()).find(o -> o.stub != null) == null;
 
     calcFakeRulesWithType();
@@ -273,7 +301,75 @@ public class ParserGenerator extends GeneratorBase {
     }
   }
 
-  @Override
+  private void openOutput(String className) throws IOException {
+    String classNameAdjusted = myPackagePrefix.isEmpty() ? className : StringUtil.trimStart(className, myPackagePrefix + ".");
+    File file = new File(myOutputPath, classNameAdjusted.replace('.', File.separatorChar) + ".java");
+    myOut = openOutputInner(className, file);
+  }
+
+  protected PrintWriter openOutputInner(String className, File file) throws IOException {
+    //noinspection ResultOfMethodCallIgnored
+    file.getParentFile().mkdirs();
+    return new PrintWriter(new FileOutputStream(file), false, this.myFile.getVirtualFile().getCharset());
+  }
+
+  private void closeOutput() {
+    myOut.close();
+  }
+
+  public void out(String s, Object... args) {
+    out(format(s, args));
+  }
+
+  public void out(String s) {
+    int length = s.length();
+    if (length == 0) {
+      myOut.println();
+      return;
+    }
+    boolean newStatement = true;
+    for (int start = 0, end; start < length; start = end + 1) {
+      boolean isComment = s.startsWith("//", start);
+      end = StringUtil.indexOf(s, '\n', start, length);
+      if (end == -1) end = length;
+      String substring = s.substring(start, end);
+      if (!isComment && (substring.startsWith("}") || substring.startsWith(")"))) {
+        myOffset--;
+        newStatement = true;
+      }
+      if (myOffset > 0) {
+        myOut.print(StringUtil.repeat("  ", newStatement ? myOffset : myOffset + 1));
+      }
+      myOut.println(substring);
+      if (isComment) {
+        newStatement = true;
+      }
+      else if (substring.endsWith("{")) {
+        myOffset++;
+        newStatement = true;
+      }
+      else if (substring.endsWith("(")) {
+        myOffset++;
+        newStatement = false;
+      }
+      else {
+        newStatement = substring.endsWith(";") || substring.endsWith("}");
+      }
+    }
+  }
+
+  public void newLine() {
+    out("");
+  }
+
+  public @NotNull String shorten(@NotNull String s) {
+    return myShortener.shorten(s);
+  }
+
+  protected void resetOffset() {
+    myOffset = 0;
+  }
+
   public void generate() throws IOException {
     {
       generateParser();
@@ -634,8 +730,17 @@ public class ParserGenerator extends GeneratorBase {
     return result;
   }
 
-  @Override
-  protected @NotNull Set<String> collectClasses(Set<String> imports, String packageName) {
+  private enum Java { CLASS, INTERFACE, ABSTRACT_CLASS }
+  private void generateClassHeader(String className,
+                                   Set<String> imports,
+                                   String annos,
+                                   Java javaType,
+                                   String... supers) {
+    generateFileHeader(className);
+    String packageName = StringUtil.getPackageName(className);
+    String shortClassName = StringUtil.getShortName(className);
+    out("package %s;", packageName);
+    newLine();
     Set<String> includedPackages = JBIterable.from(imports)
       .filter(o -> !o.startsWith("static") && o.endsWith(".*"))
       .map(o -> StringUtil.trimEnd(o, ".*"))
@@ -645,12 +750,62 @@ public class ParserGenerator extends GeneratorBase {
       if (includedPackages.contains(info.intfPackage)) includedClasses.add(StringUtil.getShortName(info.intfClass));
       if (includedPackages.contains(info.implPackage)) includedClasses.add(StringUtil.getShortName(info.implClass));
     }
-    return includedClasses;
+    NameShortener shortener = new NameShortener(packageName, !G.generateFQN);
+    shortener.addImports(imports, includedClasses);
+    for (String s : shortener.getImports()) {
+      out("import %s;", s);
+    }
+    if (G.generateFQN && imports.contains("#forced")) {
+      for (String s : JBIterable.from(imports).filter(o -> !"#forced".equals(o))) {
+        out("import %s;", s);
+      }
+    }
+    newLine();
+    StringBuilder sb = new StringBuilder();
+    for (int i = 0, supersLength = supers.length; i < supersLength; i++) {
+      String aSuper = supers[i];
+      if (StringUtil.isEmpty(aSuper)) continue;
+      if (imports.contains(aSuper + ";")) {
+        aSuper = StringUtil.getShortName(aSuper);
+      }
+      if (i == 0) {
+        sb.append(" extends ").append(shortener.shorten(aSuper));
+      }
+      else if (javaType != ParserGenerator.Java.INTERFACE && i == 1) {
+        sb.append(" implements ").append(shortener.shorten(aSuper));
+      }
+      else {
+        sb.append(", ").append(shortener.shorten(aSuper));
+      }
+    }
+    if (StringUtil.isNotEmpty(annos)) {
+      out(shortener.shorten(annos));
+    }
+    out("public %s %s%s {", Case.LOWER.apply(javaType.name()).replace('_', ' '), shortClassName, sb.toString());
+    newLine();
+    myShortener = shortener;
   }
 
-  @NotNull
-  protected String generatePackageName(String className) {
-    return StringUtil.getPackageName(className);
+  private void generateFileHeader(String className) {
+    String header = getRootAttribute(myFile, KnownAttribute.CLASS_HEADER, className);
+    String text = StringUtil.isEmpty(header) ? "" : getStringOrFile(header);
+    if (StringUtil.isNotEmpty(text)) {
+      out(text);
+    }
+    resetOffset();
+  }
+
+  private String getStringOrFile(String classHeader) {
+    try {
+      File file = new File(mySourcePath, classHeader);
+      if (file.exists()) return FileUtil.loadFile(file);
+    }
+    catch (IOException ex) {
+      LOG.error(ex);
+    }
+    return classHeader.startsWith("//") || classHeader.startsWith("/*") ? classHeader :
+           StringUtil.countNewLines(classHeader) > 0 ? "/*\n" + classHeader + "\n*/" :
+           "// " + classHeader;
   }
 
   /**
@@ -710,7 +865,7 @@ public class ParserGenerator extends GeneratorBase {
 
     String extraParameters = metaParameters.stream().map(it -> ", Parser " + it).collect(joining());
     out("%sstatic boolean %s(%s %s, int %s%s) {", !isRule ? "private " : isPrivate ? "" : "public ",
-        funcName, shorten(C.PsiBuilderClass), N.builder, N.level, extraParameters);
+        funcName, shorten(PSI_BUILDER_CLASS), N.builder, N.level, extraParameters);
     if (isSingleNode) {
       if (isPrivate && !isLeftInner && recoverWhile == null && frameName == null) {
         String nodeCall = generateNodeCall(rule, node, getNextName(funcName, 0)).render(N);
@@ -756,7 +911,7 @@ public class ParserGenerator extends GeneratorBase {
     boolean sectionMaybeDropped = sectionRequiredSimple && type == BNF_CHOICE && elementTypeRef == null &&
                                   !ContainerUtil.exists(children, o -> isRollbackRequired(o, myFile));
     String modifiers = modifierList.isEmpty()? "_NONE_" : StringUtil.join(modifierList, " | ");
-    String shortMarker = !G.generateFQN ? "Marker" : C.PsiBuilderClass + ".Marker";
+    String shortMarker = !G.generateFQN ? "Marker" : PSI_BUILDER_CLASS + ".Marker";
     if (sectionRequiredSimple) {
       if (!sectionMaybeDropped) {
         out("%s %s = enter_section_(%s);", shortMarker, N.marker, N.builder);
@@ -1364,14 +1519,14 @@ public class ParserGenerator extends GeneratorBase {
       String elementTypeClass = getAttribute(rule, KnownAttribute.ELEMENT_TYPE_CLASS);
       String elementTypeFactory = getAttribute(rule, KnownAttribute.ELEMENT_TYPE_FACTORY);
       compositeToClassAndFactoryMap.put(elementType, Trinity.create(elementTypeClass, elementTypeFactory, ruleInfo));
-      if (useFactory(elementTypeFactory)) {
+      if (elementTypeFactory != null) {
         imports.add(StringUtil.getPackageName(elementTypeFactory));
       }
       else {
         ContainerUtil.addIfNotNull(imports, elementTypeClass);
       }
     }
-    if (useFactory(tokenTypeFactory)) {
+    if (tokenTypeFactory != null) {
       imports.add(StringUtil.getPackageName(tokenTypeFactory));
     }
     else {
@@ -1396,7 +1551,7 @@ public class ParserGenerator extends GeneratorBase {
       for (String elementType : sortedCompositeTypes.keySet()) {
         Trinity<String, String, RuleInfo> info = compositeToClassAndFactoryMap.get(elementType);
         String elementCreateCall;
-        if (!useFactory(info.second)) {
+        if (info.second == null) {
           elementCreateCall = "new " + shorten(info.first);
         }
         else {
@@ -1412,7 +1567,7 @@ public class ParserGenerator extends GeneratorBase {
       String exactType = null;
       Map<String, String> sortedTokens = new TreeMap<>();
       String tokenCreateCall;
-      if (!useFactory(tokenTypeFactory)) {
+      if (tokenTypeFactory == null) {
         exactType = tokenTypeClass;
         tokenCreateCall = "new " + shorten(exactType);
       }
@@ -1509,10 +1664,6 @@ public class ParserGenerator extends GeneratorBase {
       out("}");
     }
     out("}");
-  }
-
-  protected boolean useFactory(String factory) {
-    return factory != null;
   }
 
   private boolean isIgnoredWhitespaceToken(@NotNull String tokenName, @NotNull String tokenText) {
@@ -1652,8 +1803,8 @@ public class ParserGenerator extends GeneratorBase {
       for (NavigatablePsiElement m : constructors) {
         List<String> types = myJavaHelper.getMethodTypes(m);
         Function<Integer, List<String>> annoProvider = i -> myJavaHelper.getParameterAnnotations(m, (i - 1) / 2);
-        out("public " + shortName + "(" + getParametersString(types, 1, 3, substitutor, annoProvider, getShortener()) + ") {");
-        out("super(" + getParametersString(types, 1, 2, substitutor, annoProvider, getShortener()) + ");");
+        out("public " + shortName + "(" + getParametersString(types, 1, 3, substitutor, annoProvider, myShortener) + ") {");
+        out("super(" + getParametersString(types, 1, 2, substitutor, annoProvider, myShortener) + ");");
         out("}");
         newLine();
       }
@@ -2017,7 +2168,7 @@ public class ParserGenerator extends GeneratorBase {
     }
 
     boolean many = cardinality.many();
-    String s = targetRule == null ? C.PsiElementClass : getAccessorType(targetRule);
+    String s = targetRule == null ? PSI_ELEMENT_CLASS : getAccessorType(targetRule);
     String className = shorten(s);
     String getterName = getGetterName(methodInfo.name);
     String tail = intf ? "();" : "() {";
@@ -2055,15 +2206,15 @@ public class ParserGenerator extends GeneratorBase {
     Function<String, String> substitutor = ParserGeneratorUtil::unwrapTypeArgumentForParamList;
     out("%s%s%s %s(%s)%s%s",
         intf ? "" : "public ",
-        getGenericClauseString(genericParameters, getShortener()),
+        getGenericClauseString(genericParameters, myShortener),
         returnType,
         methodName,
-        getParametersString(methodTypes, offset, 3, substitutor, annoProvider, getShortener()),
-        getThrowsString(exceptionList, getShortener()),
+        getParametersString(methodTypes, offset, 3, substitutor, annoProvider, myShortener),
+        getThrowsString(exceptionList, myShortener),
         intf ? ";" : " {");
     if (!intf) {
       String implUtilRef = shorten(StringUtil.notNullize(myPsiImplUtilClass, KnownAttribute.PSI_IMPL_UTIL_CLASS.getName()));
-      String string = getParametersString(methodTypes, offset, 2, substitutor, annoProvider, getShortener());
+      String string = getParametersString(methodTypes, offset, 2, substitutor, annoProvider, myShortener);
       out("%s%s.%s(this%s);", "void".equals(returnType) ? "" : "return ", implUtilRef, methodName,
           string.isEmpty() ? "" : ", " + string);
       out("}");
