--- testData/fleet/FleetExprParser.bnf
@@ -1,79 +1,0 @@
-/*
- * Copyright 2011-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
- */
-
-{
-  generate=[psi="no"]
-  classHeader="//header.txt"
-  parserClass="org.intellij.grammar.expression.ExpressionParser"
-  extends(".*expr")=expr
-  elementTypeFactory="org.intellij.grammar.expression.ExpressionParserDefinition.createType"
-  tokenTypeFactory="org.intellij.grammar.expression.ExpressionParserDefinition.createTokenType"
-  elementTypeHolderClass="org.intellij.grammar.expression.ExpressionTypes"
-  parserUtilClass="org.intellij.grammar.parser.GeneratedParserUtilBase"
-
-  tokens=[
-    space='regexp:\s+'
-    comment='regexp://.*'
-    number='regexp:\d+(\.\d*)?'
-    id='regexp:\p{Alpha}\w*'
-    string="regexp:('([^'\\]|\\.)*'|\"([^\"\\]|\\.)*\")"
-
-    syntax='regexp:;|\.|\+|-|\*\*|\*|==|=|/|,|\(|\)|\^|\!=|\!|>=|<=|>|<'
-  ]
-}
-root ::= element *
-private element ::= expr ';'?  {recoverWhile=element_recover}
-private element_recover ::= !('(' | '+' | '-' | '!' | 'multiply' | id | number)
-
-// left recursion and empty PSI children define expression root
-expr ::= assign_expr
-  | conditional_group
-  | add_group
-  | boolean_group
-  | mul_group
-  | unary_group
-  | exp_expr
-  | factorial_expr
-  | call_expr
-  | qualification_expr
-  | primary_group
-  {extraRoot=true}
-private boolean_group ::= xor_expr | between_expr | is_not_expr
-
-private conditional_group ::= elvis_expr | conditional_expr
-private unary_group ::= unary_plus_expr | unary_min_expr | unary_not_expr
-private mul_group ::= mul_expr | div_expr
-private add_group ::= plus_expr | minus_expr
-private primary_group ::= special_expr | simple_ref_expr | literal_expr | paren_expr
-
-// expressions: auto-operator detection or parens
-fake ref_expr ::= expr? '.' identifier
-simple_ref_expr ::= identifier {extends=ref_expr elementType=ref_expr}
-qualification_expr ::= expr '.' identifier {extends=ref_expr elementType=ref_expr}
-call_expr ::= ref_expr arg_list
-arg_list ::= '(' [ !')' expr  (',' expr) * ] ')' {pin(".*")=1}
-literal_expr ::= number
-identifier ::= id
-unary_min_expr ::= '-' expr
-unary_plus_expr ::= '+' expr
-unary_not_expr ::= '!' expr
-xor_expr ::= expr '^' expr
-assign_expr ::= expr '=' expr { rightAssociative=true }
-conditional_expr ::= expr ('<' | '>' | '<=' | '>=' | '==' | '!=') expr
-div_expr ::= expr '/' expr
-mul_expr ::= expr '*' expr
-minus_expr ::= expr '-' expr
-plus_expr ::= expr '+' expr
-exp_expr ::= expr ('**' expr) + // N-ary variant
-factorial_expr ::= expr '!'
-paren_expr ::= '(' expr ')'
-elvis_expr ::= expr '?' expr ':' expr
-is_not_expr ::= expr IS NOT expr
-between_expr ::= expr BETWEEN add_group AND add_group {
-  methods=[testExpr="expr[0]"]
-}
-
-// test specific expressions
-external special_expr ::= meta_special_expr
-meta_special_expr ::= 'multiply' '(' simple_ref_expr ',' mul_expr ')' {elementType="special_expr" pin=2}
