--- testData/fleet/FleetPsiGen.PSI.expected.java
@@ -1,45 +1,0 @@
-// ---- GeneratedTypes.java -----------------
-//header.txt
-package fleet.generated;
-
-import fleet.com.intellij.psi.tree.IElementType;
-import fleet.sample.MyRootType;
-
-public interface GeneratedTypes {
-
-  IElementType A_STATEMENT = new IElementType("A_STATEMENT", null);
-  IElementType BLOCK_OF = new IElementType("BLOCK_OF", null);
-  IElementType B_STATEMENT = new IElementType("B_STATEMENT", null);
-  IElementType CAST_EXPR = new IElementType("CAST_EXPR", null);
-  IElementType CHOICE_JOINED = new IElementType("CHOICE_JOINED", null);
-  IElementType C_STATEMENT = new IElementType("C_STATEMENT", null);
-  IElementType EXPR = new IElementType("EXPR", null);
-  IElementType GRAMMAR_ELEMENT = new IElementType("GRAMMAR_ELEMENT", null);
-  IElementType IDENTIFIER = new IElementType("IDENTIFIER", null);
-  IElementType ID_EXPR = new IElementType("ID_EXPR", null);
-  IElementType INCLUDE_SECTION = new IElementType("INCLUDE_SECTION", null);
-  IElementType INCLUDE__SECTION__ALT = new IElementType("INCLUDE__SECTION__ALT", null);
-  IElementType ITEM_EXPR = new IElementType("ITEM_EXPR", null);
-  IElementType LEFT_SHADOW = new IElementType("LEFT_SHADOW", null);
-  IElementType LEFT_SHADOW_TEST = new IElementType("LEFT_SHADOW_TEST", null);
-  IElementType LITERAL = new IElementType("LITERAL", null);
-  IElementType MISSING_EXTERNAL_TYPE = new IElementType("MISSING_EXTERNAL_TYPE", null);
-  IElementType MUL_EXPR = new IElementType("MUL_EXPR", null);
-  IElementType PLUS_EXPR = new IElementType("PLUS_EXPR", null);
-  IElementType REF_EXPR = new IElementType("REF_EXPR", null);
-  IElementType ROOT = new IElementType("ROOT", null);
-  IElementType ROOT_B = new MyRootType("ROOT_B");
-  IElementType ROOT_C = new MyRootType("ROOT_C");
-  IElementType ROOT_D = new MyRootType("ROOT_D");
-  IElementType SOME_EXPR = new IElementType("SOME_EXPR", null);
-  IElementType SPECIAL_REF = new IElementType("SPECIAL_REF", null);
-  IElementType STATEMENT = new IElementType("STATEMENT", null);
-
-  IElementType ID = new IElementType("id", null);
-  IElementType NOTSPACE = new IElementType("notspace", null);
-  IElementType NUMBER = new IElementType("number", null);
-  IElementType OF = new IElementType("OF", null);
-  IElementType OP_DIV = new IElementType("/", null);
-  IElementType OP_MUL = new IElementType("*", null);
-  IElementType SLASH = new IElementType("\\", null);
-}
