--- testData/fleet/IFileTypeGeneration.bnf
@@ -1,9 +1,0 @@
-{
-  extraRoot(".*")=true
-}
-
-grammar ::= list | map
-list ::= '(' element (',' element) * ')'
-map ::= '(' entry (',' entry) * ')'
-entry ::= 'name' '->' element
-element ::= 'id'
