+++ tests/org/intellij/grammar/BnfTestSuite.java
@@ -28,7 +28,6 @@ public class BnfTestSuite extends TestCase {
       testSuite.addTestSuite(JFlexParserTest.class);
       testSuite.addTestSuite(BnfParserTest.class);
       testSuite.addTestSuite(BnfGeneratorTest.class);
-      testSuite.addTestSuite(FleetBnfGeneratorTest.class);
       testSuite.addTestSuite(ExpressionParserTest.class);
       testSuite.addTestSuite(BnfLivePreviewParserTest.class);
       return testSuite;
